Move CustomError to error module; it's no longer needed
Ideally I should remove the code, will do that in a final pass. Consider
this a rejuggling of furniture first.
