+++ Cargo.toml
@@ -39,7 +39,7 @@ serde_json = "1.0.87"
 heapless = { version = "0.7.16", default-features = false }
 cstr_core = "0.2.6"
 uncased = { version = "0.9.7" }
-shared-bus = "0.2.4"
+# shared-bus = "0.2.4" # No longer compatible
 once_cell = "1.16.0"
 
 # time = { version = "0.3.17", features = ["macros"] }
