+++ src/error.rs
@@ -67,3 +67,59 @@ impl From<anyhow::Error> for BlanketError {
         }
     }
 }
+
+// CustomError
+// NOTE: This is temporarily being stashed away here, if I need a template for a NewType wrapper.
+#[derive(Debug)]
+pub struct CustomError(Option<BoxError>);
+
+impl CustomError {
+    /// Create a new `Error` from a boxable error.
+    pub fn new() -> Self {
+        Self(None)
+    }
+
+    pub fn set_error(&mut self, error: impl Into<BoxError>) {
+        self.0 = Some(error.into());
+    }
+
+    /// Convert an `Error` back into the underlying boxed trait object.
+    pub fn into_inner(self) -> BoxError {
+        self.0.expect("Unwrapping option for CustomError")
+    }
+}
+
+impl fmt::Display for CustomError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.0
+            .as_ref()
+            .expect("Unwrapping option for CustomError")
+            .fmt(f)
+    }
+}
+
+impl StdError for CustomError {
+    fn source(&self) -> Option<&(dyn StdError + 'static)> {
+        Some(
+            &*self
+                .0
+                .as_deref()
+                .expect("Unwrapping option for CustomError"),
+        )
+    }
+}
+
+impl From<anyhow::Error> for CustomError {
+    fn from(error: anyhow::Error) -> Self {
+        let mut e = Self::new();
+        e.set_error(error);
+
+        e
+    }
+}
+
+impl From<CustomError> for BlanketError {
+    fn from(error: CustomError) -> Self {
+        Self::new(error)
+    }
+}
